requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0.1
google-generativeai>=0.8.0
PyPDF2>=3.0.1
python-docx>=1.1.2
reportlab>=4.2.0
